pyproject.toml
[project]
name = "august"
version = "0.1.0"
description = "Your daily brief, powered by AI - 热点·摘要·播报"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.2",
    "httpx>=0.28.1",
    "uvicorn[standard]>=0.30.0",
    "apscheduler>=3.10.4",
    "litellm>=1.80.0",
    "python-dotenv>=1.2.1",
    "edge-tts>=7.2.3",
    "feedparser>=6.0.11",
]
